Not much progress, commit before creating branches to debug latency

+++ osc.py
@@ -29,7 +29,7 @@ class osc():
     def generateWavedata(self, n_samples: int = consts.BUFFER_SIZE, frequency: float = 200.00, MIDI: int = 0) -> list:
         
         #Create enough samples to fill the requested buffer size
-        samples = np.zeros(n_samples, dtype=np.float32)
+        samples = np.zeros(n_samples, dtype=np.float64)
 
         #NOT the position as used for audio output, this tracks generation progress
         generation_phase = 0.000
